+++ src/providers/authProvider/index.ts
@@ -4,7 +4,8 @@ import {
   type DataProvider
 } from 'react-admin'
 import * as constants from '../../constants'
-import { trackEvent, AuditType } from '../../utils/audit'
+import { trackEvent } from '../../utils/audit'
+import { AuditType } from '../../utils/activity-types'
 import {
   decryptData,
   decryptPassword,
